# openapiconfig-privsense.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "PrivSense API",
      "description": "API for detecting personally identifiable information (PII) in relational databases",
      "contact": {
        "name": "PrivSense Team",
        "url": "https://www.privsense.com",
        "email": "support@privsense.com"
      },
      "license": {
        "name": "Proprietary",
        "url": "https://www.privsense.com/license"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/privsense",
        "description": "Default server URL"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
...
    "paths": {
...
      "/api/v1/dashboard/trends": {
        "get": {
          "tags": [
            "Dashboard"
          ],
          "summary": "Get scan trends data",
          "description": "Returns time series data for chart visualizations",
          "operationId": "getScanTrends",
          "parameters": [
            {
              "name": "timeframe",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "default": "30"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Trends data retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTrendsDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/dashboard/top-findings": {
        "get": {
          "tags": [
            "Dashboard"
          ],
          "summary": "Get top PII findings",
          "description": "Returns the most common PII types found",
          "operationId": "getTopPiiFindings",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 5
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Top findings retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PiiSummaryDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/dashboard/recent-scans": {
        "get": {
          "tags": [
            "Dashboard"
          ],
          "summary": "Get recent scans",
          "description": "Returns the most recent scan jobs",
          "operationId": "getRecentScans",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 5
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Recent scans retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Link": {
          "type": "object",
          "properties": {
            "href": {
              "type": "string"
            },
            "hreflang": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "deprecation": {
              "type": "string"
            },
            "profile": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "templated": {
              "type": "boolean"
            }
          }
        },
        "ScanJobResponse": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "jobId": {
              "type": "string",
              "format": "uuid"
            },
            "connectionId": {
              "type": "string",
              "format": "uuid"
            },
            "status": {
              "type": "string"
            },
            "currentOperation": {
              "type": "string"
            },
            "startTime": {
              "type": "string",
              "format": "date-time"
            },
            "endTime": {
              "type": "string",
              "format": "date-time"
            },
            "lastUpdateTime": {
              "type": "string",
              "format": "date-time"
            },
            "progress": {
              "type": "integer",
              "format": "int32"
            },
            "completed": {
              "type": "boolean"
            },
            "failed": {
              "type": "boolean"
            },
            "errorMessage": {
              "type": "string"
            },
            "databaseName": {
              "type": "string"
            },
            "databaseProductName": {
              "type": "string"
            },
            "totalColumnsScanned": {
              "type": "integer",
              "format": "int32"
            },
            "totalPiiColumnsFound": {
              "type": "integer",
              "format": "int32"
            },
            "success": {
              "type": "boolean"
            }
          }
        },
...
        "ScanTrendsDTO": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "timeframe": {
              "type": "string"
            },
            "startDate": {
              "type": "string"
            },
            "endDate": {
              "type": "string"
            },
            "scanCountByDate": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int64"
              }
            },
            "piiCountByDate": {
              "type": "object",
              "additionalProperties": {
                "type": "integer",
                "format": "int64"
              }
            },
            "timeSeriesData": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeSeriesDataPoint"
              }
            },
            "success": {
              "type": "boolean"
            }
          }
        },
        "TimeSeriesDataPoint": {
          "type": "object",
          "properties": {
            "date": {
              "type": "string",
              "format": "date"
            },
            "scanCount": {
              "type": "integer",
              "format": "int64"
            },
            "piiCount": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "PiiSummaryDTO": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "piiType": {
              "type": "string"
            },
            "count": {
              "type": "integer",
              "format": "int32"
            },
            "percentage": {
              "type": "number",
              "format": "double"
            },
            "success": {
              "type": "boolean"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "Enter JWT token with Bearer prefix. Example: Bearer eyJhbGciOiJIUzI1NiIsIn...",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }